requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt -
#
-e file:///Users/clpezma/PycharmProjects/tm
argon2-cffi==18.3.0
arrow==0.12.1
authomatic==0.1.0.post1
certifi==2018.10.15       # via requests
cffi==1.11.5              # via argon2-cffi
chardet==3.0.4            # via requests
cornice==3.4.2
cssselect==1.0.3          # via premailer
cssutils==1.0.2           # via premailer
hupper==1.4.1             # via pyramid
idna==2.7                 # via requests
jinja2==2.10              # via pyramid-jinja2
lxml==4.2.5               # via premailer
mako==1.0.7               # via pyramid-mako
markupsafe==1.1.0         # via jinja2, mako, pyramid-jinja2
marshmallow==v3.0.0b20
pastedeploy==1.5.2        # via plaster-pastedeploy
plaster-pastedeploy==0.6  # via pyramid
plaster==1.0              # via plaster-pastedeploy, pyramid
premailer==3.2.0
psycopg2==2.7.6.1
pycparser==2.19           # via cffi
pygments==2.2.0           # via pyramid-debugtoolbar
pyjwt==1.6.4              # via pyramid-jwt
pyramid-debugtoolbar==4.5
pyramid-jinja2==2.7
pyramid-jwt==1.4.1
pyramid-mailer==0.15.1
pyramid-mako==1.0.2       # via pyramid-debugtoolbar
pyramid-retry==1.0
pyramid-tm==2.2.1
pyramid==1.10.1           # via cornice, pyramid-debugtoolbar, pyramid-jinja2, pyramid-jwt, pyramid-mailer, pyramid-mako, pyramid-retry, pyramid-tm
python-dateutil==2.7.5    # via arrow
pytz==2018.7
repoze.lru==0.7           # via pyramid-debugtoolbar
repoze.sendmail==4.4.1    # via pyramid-mailer
requests==2.20.1          # via premailer
simplejson==3.16.0        # via cornice
six==1.11.0               # via argon2-cffi, cornice, python-dateutil, sqlalchemy-utils
sqlalchemy-utils==0.33.8
sqlalchemy==1.2.14        # via sqlalchemy-utils, zope.sqlalchemy
transaction==2.4.0        # via pyramid-mailer, pyramid-tm, repoze.sendmail, zope.sqlalchemy
translationstring==1.3    # via pyramid
urllib3==1.24.1           # via requests
venusian==1.1.0           # via cornice, pyramid
waitress==1.1.0
webob==1.8.4              # via pyramid
zope.deprecation==4.3.0   # via pyramid, pyramid-jinja2
zope.interface==4.6.0     # via pyramid, pyramid-retry, repoze.sendmail, transaction, zope.sqlalchemy
zope.sqlalchemy==1.0
